# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Employee Management System API",
        "description": "This is the API documentation for the Employee Management System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "description": "Fetch all employees",
                "operationId": "ede932cb6e7161a5520a90901e3a1270",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/PageNumber"
                    },
                    {
                        "$ref": "#/components/parameters/PageLimit"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employees fetched successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetAllEmployeesResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "description": "Create an employee",
                "operationId": "a5af921aac4bf43d4d5e7caeeea41795",
                "requestBody": {
                    "description": "Employee details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreEmployeeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/StoreEmployeeSuccess"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "422": {
                        "$ref": "#/components/responses/StoreEmployeeRequestValidation"
                    }
                }
            }
        },
        "/api/employees/{id}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "description": "Fetch an employee",
                "operationId": "3d4148e9afa2fa52d618e8935f33ed83",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/EmployeeId"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/EmployeeFetchSuccess"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    }
                }
            },
            "put": {
                "tags": [
                    "Employees"
                ],
                "summary": "Update an (PUT/PATCH)",
                "description": "Update an employee by ID. This endpoint supports both PUT and PATCH requests.",
                "operationId": "55e1ceec679a0d1436c869735e984a83",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/EmployeeId"
                    }
                ],
                "requestBody": {
                    "description": "Employee details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEmployeeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/EmployeeFetchSuccess"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "422": {
                        "$ref": "#/components/responses/StoreEmployeeRequestValidation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Employees"
                ],
                "description": "Delete an employee",
                "operationId": "00c3138566163aa6d741a37971022e5c",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/EmployeeId"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/DeleteEmployeeSuccess"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Employee": {
                "title": "Employee",
                "description": "Employee schema",
                "required": [
                    "names",
                    "email",
                    "phone_number",
                    "badge_id"
                ],
                "properties": {
                    "id": {
                        "description": "The id of the employee",
                        "type": "integer"
                    },
                    "names": {
                        "description": "The names of the employee",
                        "type": "string"
                    },
                    "email": {
                        "description": "The email of the employee",
                        "type": "string",
                        "format": "email"
                    },
                    "phone_number": {
                        "description": "The phone number of the employee",
                        "type": "string"
                    },
                    "badge_id": {
                        "description": "The badge id of the employee",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "The date and time the employee was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The date and time the employee was updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "description": "The date and time the employee was deleted",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "example": {
                    "names": "Sheja Eddy",
                    "email": "sheja@eddy.com",
                    "phone_number": "250784141587",
                    "badge_id": "SH123456"
                }
            },
...
            "GetAllEmployeesResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 200
                    },
                    "message": {
                        "type": "string",
                        "example": "Employees fetched successfully"
                    },
                    "data": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "data": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            },
                            "first_page_url": {
                                "type": "string",
                                "example": "http://localhost:8000/api/employees?page=1"
                            },
                            "from": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 2
                            },
                            "last_page_url": {
                                "type": "string",
                                "example": "http://localhost:8000/api/employees?page=2"
                            },
                            "next_page_url": {
                                "type": "string",
                                "example": "http://localhost:8000/api/employees?page=2"
                            },
                            "path": {
                                "type": "string",
                                "example": "http://localhost:8000/api/employees"
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 5
                            },
                            "prev_page_url": {
                                "type": "string",
                                "example": "http://localhost:8000/api/employees?page=1"
                            },
                            "to": {
                                "type": "integer",
                                "example": 5
                            },
                            "total": {
                                "type": "integer",
                                "example": 6
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "StoreEmployeeRequest": {
                "required": [
                    "names",
                    "email",
                    "badge_id",
                    "phone_number"
                ],
                "properties": {
                    "names": {
                        "description": "The names of the employee",
                        "type": "string",
                        "example": "Sheja Eddy"
                    },
                    "email": {
                        "description": "The email of the employee",
                        "type": "string",
                        "format": "email",
                        "example": "sheja@eddy.com"
                    },
                    "badge_id": {
                        "description": "The badge id of the employee",
                        "type": "string",
                        "example": "SH123456"
                    },
                    "phone_number": {
                        "description": "The phone number of the employee",
                        "type": "string",
                        "example": "250784141587"
                    }
                },
                "type": "object"
            },
            "UpdateEmployeeRequest": {
                "required": [],
                "properties": {
                    "names": {
                        "description": "The names of the employee",
                        "type": "string",
                        "example": "Sheja Eddy"
                    },
                    "email": {
                        "description": "The email of the employee",
                        "type": "string",
                        "format": "email",
                        "example": "sheja@eddy.com"
                    },
                    "badge_id": {
                        "description": "The badge id of the employee",
                        "type": "string",
                        "example": "SH123456"
                    },
                    "phone_number": {
                        "description": "The phone number of the employee",
                        "type": "string",
                        "example": "250784141587"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "DeleteEmployeeSuccess": {
                "description": "Employee deleted successfully",
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "status",
                                "message"
                            ],
                            "properties": {
                                "status": {
                                    "description": "The status of the response",
                                    "type": "integer",
                                    "example": 200
                                },
                                "message": {
                                    "description": "The message of the response",
                                    "type": "string",
                                    "example": "Employee deleted successfully"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "EmployeeFetchSuccess": {
                "description": "Employee details fetched successfully",
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "status",
                                "message",
                                "data"
                            ],
                            "properties": {
                                "status": {
                                    "description": "The status of the response",
                                    "type": "integer",
                                    "example": 200
                                },
                                "message": {
                                    "description": "The message of the response",
                                    "type": "string",
                                    "example": "Employee details fetched successfully"
                                },
                                "data": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "InternalServerError": {
                "description": "Internal server error",
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "status",
                                "message"
                            ],
                            "properties": {
                                "status": {
                                    "description": "The status of the response",
                                    "type": "integer",
                                    "example": 500
                                },
                                "message": {
                                    "description": "The message of the response",
                                    "type": "string",
                                    "example": "Internal server error"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "NotFound": {
                "description": "Resource not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "status",
                                "message"
                            ],
                            "properties": {
                                "status": {
                                    "description": "The status of the response",
                                    "type": "integer",
                                    "example": 404
                                },
                                "message": {
                                    "description": "The message of the response",
                                    "type": "string",
                                    "example": "Resource not found"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "StoreEmployeeRequestValidation": {
                "description": "Validation failed",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "example": 422
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Validation failed"
                                },
                                "errors": {
                                    "properties": {
                                        "names": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The names field is required"
                                            }
                                        },
                                        "email": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The email address is already in use"
                                            }
                                        },
                                        "badge_id": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The badge ID is already in use"
                                            }
                                        },
                                        "phone_number": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The phone number field is already in use"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "StoreEmployeeSuccess": {
                "description": "Employee created successfully",
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "status",
                                "message",
                                "data"
                            ],
                            "properties": {
                                "status": {
                                    "description": "The status of the response",
                                    "type": "integer",
                                    "example": 201
                                },
                                "message": {
                                    "description": "The message of the response",
                                    "type": "string",
                                    "example": "Employee created successfully"
                                },
                                "data": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Unauthenticated": {
                "description": "Unauthorized access",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "example": 401
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Unauthorized access, please login."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "parameters": {
            "PageLimit": {
                "name": "PageLimit",
                "in": "query",
                "description": "The number of items per page",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 10
                }
            },
            "PageNumber": {
                "name": "PageNumber",
                "in": "query",
                "description": "The page number",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 1
                }
            },
...
            "EmployeeId": {
                "name": "EmployeeId",
                "in": "path",
                "description": "Employee id parameter",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    }
}